pyproject.toml
[tool.poetry]
name = "xpulumi"
version = "1.0.0"
description = "Tools to make Pulumi more usable"
authors = ["Sam McKelvie <dev@mckelvie.org>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
colorama = "^0.4.4"
jq = "^1.2.2"
argcomplete = "^2.0.0"
pulumi = "^3.26.1"
boto3 = "^1.21.18"
secret-kv = "^1.3.0"
pycryptodomex = "^3.14.1"
requests = "^2.27.1"

mypy = "^0.931"
python-semantic-release = "^7.25.2"
boto3-stubs = {extras = ["essential"], version = "^1.21.34"}
types-requests = "^2.27.12"
pycryptodome-test-vectors = "^1.0.7"
types-PyYAML = "^6.0.5"
importlib-resources = "^5.4.0"
packaging = "^21.3"
xdg = "^5.1.1"
"ruamel.yaml" = "^0.17.21"
pulumi-aws = "^4.38.1"
boto3-type-annotations = "^0.3.1"
debugpy = "^1.6.0"
yq = "^2.14.0"
cloud-init-gen = "^1.1.0"
#project-init-tools = { git = "https://github.com/sammck/project-init-tools.git", branch = "main" }
project-init-tools = { file = "../project-init-tools/dist/project-init-tools-0.1.0.tar.gz" }
#project-init-tools = { file = "../project-init-tools" }
#project-init-tools = "../project-init-tools"

[tool.poetry.dev-dependencies]
pylint = "^2.13.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
xpulumi = 'xpulumi.__main__:run'
pulumi = 'xpulumi.pulumi_cli.__main__:run'

[tool.semantic_release]
version_variable = 'xpulumi/version.py:__version__'
version_toml = 'pyproject.toml:tool.poetry.version'
branch = 'main'
upload_to_pypi = false
upload_to_release = true
build_command = "pip install poetry && poetry build"

[tool.pylint.'MESSAGES CONTROL']
disable=[
    "wrong-import-order",
    "duplicate-code",
    "too-many-arguments",
    "missing-function-docstring",
    "import-outside-toplevel",
    "too-few-public-methods",
    "missing-class-docstring",
    "unused-import",
    "too-many-locals",
    "unused-argument",
    "invalid-name",
    "no-self-use",
    "global-statement",
    "broad-except",
    "too-many-branches",
    "too-many-statements",
    "exec-used",
    "ungrouped-imports",
    "subprocess-popen-preexec-fn",
    "multiple-statements",
    "too-many-public-methods",
    "missing-module-docstring",
    "too-many-instance-attributes",
    "too-many-nested-blocks",
    "unneeded-not",
    "unnecessary-lambda",
  ]

[tool.pylint.MASTER]
ignore=[
    ".venv",
    "trash"
  ]

[tool.pylint.FORMAT]
indent-after-paren=4
indent-string='  '
max-line-length=200
